+++ CHANGES.md
@@ -17,6 +17,9 @@
   - Fix `Tree.fold ~force:(False f)` where results where partially skipped
     (#1174, @Ngoguey42, @samoht and @CraigFe )
 
+  - Fix `Tree.kind`. Empty path on a tree used to return a None instead of a
+	`` `Node``. (#1218, @Ngoguey42)
+
 - **ppx_irmin**
   - Fix a bug causing certain type derivations to be incorrect due to unsound
     namespacing. (#1083, @CraigFe)